requirements.txt
pandas
numpy
appnope==0.1.4
asttokens==2.4.1
asyncio==3.4.3
backcall==0.2.0
backoff==2.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coinbase-advanced-py==1.4.1
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
gemini-api==0.1.6
idna==3.7
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jupyter-client==8.6.2
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-kraken-sdk==1.6.2
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
zipp==3.19.2
python-dotenv